lingering dtype test issue

+++ dask/array/tests/test_array_core.py
@@ -517,7 +517,7 @@ def test_map_blocks2():
         return np.ones_like(block) * sum(block_id)
 
     d = d.map_blocks(func, dtype='i8')
-    expected = np.array([0, 0, 1, 1, 2, 2, 3, 3, 4, 4])
+    expected = np.array([0, 0, 1, 1, 2, 2, 3, 3, 4, 4], dtype='i8')
 
     assert eq(d, expected)
